Create a new BitwiseMapOp abstract base class for bitwise operators
 - for issue #308
