+++ mrgeo-mapalgebra/mrgeo-mapalgebra-rastermath/src/main/scala/org/mrgeo/mapalgebra/binarymath/BitwiseOrMapOp.scala
@@ -21,9 +21,7 @@ object BitwiseOrMapOp extends MapOpRegistrar {
     new BitwiseOrMapOp(node, variables)
 }
 
-class BitwiseOrMapOp extends RawBinaryMathMapOp {
-
-  val EPSILON: Double = 1e-12
+class BitwiseOrMapOp extends BitwiseMapOp {
 
   private[binarymath] def this(raster: Option[RasterMapOp], paramB:Option[Any]) = {
     this()
@@ -45,46 +43,6 @@ class BitwiseOrMapOp extends RawBinaryMathMapOp {
     this()
 
     initialize(node, variables)
-  }
-
-  override def execute(context: SparkContext): Boolean = {
-    varA match {
-      case Some(a) => {
-        val metadata = a.metadata().getOrElse(throw new ParserException("Uh oh - no metadata available"))
-        if (RasterUtils.isFloatingPoint(metadata.getTileType)) {
-          log.warn("Using a floating point raster like " + metadata.getPyramid + " is not recommended for bitwise operators")
-        }
-      }
-      case None => {}
-    }
-    constA match {
-      case Some(a) => {
-        val remainder = a % 1
-        if (remainder < -EPSILON || remainder > EPSILON) {
-          log.warn("Using floating point values in bitwise operators is not recommended")
-        }
-      }
-      case None => {}
-    }
-    varB match {
-      case Some(b) => {
-        val metadata = b.metadata().getOrElse(throw new ParserException("Uh oh - no metadata available"))
-        if (RasterUtils.isFloatingPoint(metadata.getTileType)) {
-          log.warn("Using a floating point raster like " + metadata.getPyramid + " is not recommended for bitwise operators")
-        }
-      }
-      case None => {}
-    }
-    constB match {
-      case Some(b) => {
-        val remainder = b % 1
-        if (remainder < -EPSILON || remainder > EPSILON) {
-          log.warn("Using floating point values in bitwise operators is not recommended")
-        }
-      }
-      case None => {}
-    }
-    return super.execute(context)
   }
 
   override private[binarymath] def function(a: Double, b: Double): Double = {
